+++ dbms/src/TestUtils/MPPTaskTestUtils.cpp
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 #include <Debug/dbgFuncCoprocessor.h>
+#include <Debug/dbgFuncCoprocessorUtils.h>
 #include <Server/MockComputeClient.h>
 #include <TestUtils/MPPTaskTestUtils.h>
